\name{getModel}
\alias{getModel}
\title{Get the parameters of a model}
\description{Generic method to get the parameters of a model}
\usage{getModel( object, \ldots)}
\arguments{
  \item{object}{an object representing a model}
  \item{...}{ adding arguments (depending on the object type)}
}
\value{
  Return parameters of the model.
}

\keyword{graphs}
\keyword{cluster}
