\name{getAdjacencyMatrix-methods}
\docType{methods}
\alias{getAdjacencyMatrix-methods}
\alias{getAdjacencyMatrix,spmgraph-method}
\title{Methods for Function `getAdjacencyMatrix' in package `spmgraph'}
\description{
	Return the adjacency matrix of an object of the `spmgraph' class.
}
\section{Methods}{
\describe{
\item{g = "spmgraph"}{ an object of the `spmgraph' class. }
\item{withlabels = FALSE}{ of set to true, the rows and columns of the matrix will be named after the graph node labels.}
}}

\keyword{methods}

