\name{getEdges-methods}
\docType{methods}
\alias{getEdges-methods}
\alias{getEdges,spmgraph-method}
\title{Methods for Function getEdges in Package `spmgraph'}
\description{
	Return the edges of an object of class `spmgraph'.
}
\section{Methods}{
\describe{
\item{g = "spmgraph"}{ an object of the `spmgraph' class. }
\item{withlabels = FALSE}{ if TRUE, the contents of the two-line matrix are the original node labels instead of their indices.}
}}
\keyword{methods}

