\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,spmgraph-method}
\title{ Methods for Function initialize in Package `spmgraph' }
\description{
  Build a new 'spmgraph' object.
}
\section{Methods}{
\describe{
\item{.Object = "spmgraph"}{ the character vector with the class name.}
}}
\keyword{methods}

