% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXER.R
\name{MIXER}
\alias{MIXER}
\title{MIXER}
\usage{
MIXER(
  y_train,
  feature_train,
  y_val,
  feature_val,
  y_test,
  feature_test,
  top_k = 5000,
  family = "binomial",
  nfolds_ridge = 20,
  n_threshold = 2000,
  min_num = 10,
  max_prop = 0.05,
  lambda_init_min = 150,
  lambda_init_max = 200,
  nlambda_adalasso = 100,
  threshold_list = NULL,
  n_grid_eval = 1000,
  threshold_grid = c("prob_range", "unit")
)
}
\arguments{
\item{y_train}{Training outcome vector.}

\item{feature_train}{Training feature matrix/data.frame (samples x features).}

\item{y_val}{Validation outcome vector.}

\item{feature_val}{Validation feature matrix/data.frame (samples x features).}

\item{y_test}{Test outcome vector.}

\item{feature_test}{Test feature matrix/data.frame (samples x features).}

\item{top_k}{Number of top features per metric used for ridge regression.}

\item{family}{Model family passed to glmnet (default: "binomial").}

\item{nfolds_ridge}{Number of CV folds for ridge regression.}

\item{n_threshold}{Number of thresholds for PIM construction.}

\item{min_num}{Minimum target feature count for lambda_max tuning.}

\item{max_prop}{Maximum proportion of features allowed for lambda_min tuning.}

\item{lambda_init_min}{Initial guess for lambda_min.}

\item{lambda_init_max}{Initial guess for lambda_max.}

\item{nlambda_adalasso}{Number of lambdas for adaptive LASSO CV grid.}

\item{threshold_list}{Optional numeric vector of thresholds used in test evaluation.
If NULL, thresholds are constructed automatically.}

\item{n_grid_eval}{Number of thresholds if thresholds are constructed automatically (default 1000).}

\item{threshold_grid}{One of c("prob_range","unit"). If \code{"prob_range"}, thresholds are
generated from the predicted-probability range; if \code{"unit"}, thresholds use a fixed
grid on \eqn{[0,1]}.}
}
\value{
A list containing ranked features, ridge coefficients, PIM, feature weights,
adaptive LASSO selected features, and test-set performance of the final model.
}
\description{
Multi-metric Integration for eXplanatory and prEdictive Ranking.
}
